# packages/backend/docs/openAPISchema.yaml
openapi: 3.0.0
paths:
  /auth/login:
    post:
      operationId: login
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
      tags: &a1
        - auth
  /auth/profile:
    get:
      operationId: profile
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProfileResponse'
      tags: *a1
  /entry/new:
    post:
      operationId: createEntry
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntryDTO'
      responses:
        '201':
          description: ''
      tags: &a2
        - entry
  '/entry/{entryId}':
    get:
      operationId: find
      parameters:
        - name: entryId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *a2
  /entry/search:
    post:
      operationId: search
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchEntryDTO'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
      tags: *a2
info:
  title: Bright-Memory-API
  description: Bright-Memory Backend service documents
  version: 0.0.1
  contact: {}
tags: []
servers: []
components:
  schemas:
    LoginResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        accessToken:
          type: string
      required:
        - id
        - name
        - accessToken
    GetProfileResponse:
      type: object
      properties:
        _id:
          type: string
          description: for internal management
        screenName:
          type: string
        loginName:
          type: string
        version:
          type: number
        createdAt:
          type: string
          description: ISO String Date
        updatedAt:
          type: string
          description: ISO String Date
      required:
        - _id
        - screenName
        - loginName
        - version
        - createdAt
        - updatedAt
    CreateEntryDTO:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    sortOperators:
      type: string
      enum:
        - createdAt
        - updatedAt
    orderOperators:
      type: string
      enum:
        - desc
        - asc
    Pagination:
      type: object
      properties:
        pageNumber:
          type: number
        size:
          type: number
        sort:
          $ref: '#/components/schemas/sortOperators'
        order:
          $ref: '#/components/schemas/orderOperators'
      required:
        - pageNumber
        - size
    SearchEntryDTO:
      type: object
      properties:
        title:
          type: string
        note:
          type: string
        url:
          type: string
        isFavorite:
          type: boolean
        pagination:
          $ref: '#/components/schemas/Pagination'
    Entry:
      type: object
      properties:
        _id:
          type: string
          description: for internal management
        owner:
          type: string
        url:
          type: string
        title:
          type: string
        note:
          type: string
        ogpImageLink:
          type: string
        isFavorite:
          type: boolean
        version:
          type: number
        createdAt:
          type: string
          description: ISO String Date
        updatedAt:
          type: string
          description: ISO String Date
      required:
        - _id
        - owner
        - url
        - title
        - note
        - ogpImageLink
        - isFavorite
        - version
        - createdAt
        - updatedAt